Slight refactor to clean things up.
Still not ideal, but it's a bit better.
